esy: init at 0.5.8
Aborted attempt to add the esy package manager to nixpkgs.  I think I
was almost there to get its dependencies, but I have no idea if
building it like a dune package would have been enough to make it
work.  The last hurdle has been with cudf and dose3.  The first
problem when building dose3 was a string-byte problem, fixed by
importing patches from the opam repository.  I don't know how to fix the
next one: dose3 attempts to find a cudf.cmxa in a folder that contains
a cudf.cma but no cmxa.

This submission is in the hope that this PR can be partially salvaged
and my work was not totally useless.